requirements.txt
streamlit>=1.32.0
anthropic>=0.25.0
scikit-learn>=1.4.0
pydantic>=2.6.0
python-dotenv>=1.0.0
pytest>=8.0.0
pandas>=2.2.0
numpy>=1.26.0
